rename dirs and files
