+++ .vscode/launch.json
@@ -1,8 +1,6 @@
 {
     "version": "0.2.0",
     "configurations": [
-        
-    
         {
             "name": "(gdb) Launch",
             "type": "cppdbg",
