+++ CMakeLists.txt
@@ -2,6 +2,7 @@ cmake_minimum_required (VERSION 3.21.0)
 project (tftp_agent)
 set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED True)
+set(CMAKE_BUILD_TYPE Debug)
 enable_testing()
 add_subdirectory(src)
 add_subdirectory(test)
