+++ src/Agent/main.cpp
@@ -1,4 +1,4 @@
-#include "tftp_client.h"
+#include "TFTPClient.h"
 #include <algorithm>
 #include <chrono>
