+++ src/CMakeLists.txt
@@ -1,14 +1,11 @@
-include_directories(transport
-                    transport/udp_socket
+include_directories(Transport
+                   # transport/udp_socket
 )
-add_executable (agent
-                transport/udp_socket/udp_socket.cpp
-                transport/tftp_client.cpp
-                agent/main.cpp
+add_executable (Agent
+                Transport/UDPSocket/UDPSocket.cpp
+                Transport/TFTPClient.cpp
+                Agent/main.cpp
 )
 
-     
-
-
-add_library(udp_socket transport/udp_socket/udp_socket.cpp)
-add_library(tftp_client transport/tftp_client.cpp)
+add_library(udp_socket Transport/UDPSocket/UDPSocket.cpp)
+add_library(tftp_client Transport/TFTPClient.cpp)
